refactor: move common classes to model module
