+++ src/pdf2aas/extractor/__init__.py
@@ -1,12 +1,10 @@
 """Module containing different extractors for the PDF2AAS workflow."""
-from .property import Property
 from .core import Extractor
 from .customLLMClient import CustomLLMClient, CustomLLMClientHTTP
 from .propertyLLM import PropertyLLM
 from .propertyLLMSearch import PropertyLLMSearch
 
 __all__ = [
-    "Property",
     "Extractor",
     "PropertyLLM",
     "PropertyLLMSearch",
