+++ src/pdf2aas/model/__init__.py
@@ -1,0 +1,11 @@
+"""Classes to handle properties their definitions and classes inside the library."""
+
+from .propertyDefinition import PropertyDefinition
+from .classDefinition import ClassDefinition
+from .property import Property
+
+__all__ = [
+    "PropertyDefinition",
+    "ClassDefinition",
+    "Property",
+]
