+++ src/pdf2aas/model/classDefinition.py
