+++ src/pdf2aas/model/propertyDefinition.py
